Basic Working Item
It works! This program can be run through the command console with node js (it requires a few dependencies). To run, use the command "node app.js" in the directory where the files are located.

Any user can connect with the localhost on port 8080 (or whatever port you're using, it tells you) or your ipv4 address.

Usernames must be between 3 and 32 characters in length. Messages must be between 1 and 1024 characters in length. There are some commands which you can see by typing "/help" in the chat (without the quotes).